# Optimize DR lookup in EDS