--- a/releasenotes/25519.yaml
+++ b/releasenotes/25519.yaml
@@ -1,7 +1,0 @@
-apiVersion: release-notes/v1
-kind: bug-fix
-area: telemetry
-issue: 25154
-releaseNotes: |
-  Removed the `pilot_xds_eds_instances` and `pilot_xds_eds_all_locality_endpoints` Istiod metrics, which were not
-  accurate.