```
name                         old time/op    new time/op    delta
EndpointGeneration/1/100-8      257µs ±16%     157µs ±11%  -39.00%  (p=0.008 n=5+5)
EndpointGeneration/10/10-8      100µs ± 9%     102µs ±14%     ~     (p=0.841 n=5+5)
EndpointGeneration/100/10-8     590µs ±13%     626µs ±11%     ~     (p=0.310 n=5+5)
EndpointGeneration/1000/1-8     869µs ± 8%     940µs ± 7%     ~     (p=0.095 n=5+5)

name                         old alloc/op   new alloc/op   delta
EndpointGeneration/1/100-8     75.1kB ± 0%    75.1kB ± 0%     ~     (p=0.357 n=5+5)
EndpointGeneration/10/10-8     22.2kB ± 0%    22.2kB ± 0%     ~     (p=0.095 n=5+5)
EndpointGeneration/100/10-8    41.9kB ± 0%    42.0kB ± 0%     ~     (p=0.730 n=5+5)
EndpointGeneration/1000/1-8    37.7kB ± 3%    37.9kB ± 1%     ~     (p=0.690 n=5+5)

name                         old allocs/op  new allocs/op  delta
EndpointGeneration/1/100-8      1.30k ± 0%     1.30k ± 0%     ~     (all equal)
EndpointGeneration/10/10-8        395 ± 0%       395 ± 0%     ~     (all equal)
EndpointGeneration/100/10-8       401 ± 1%       401 ± 0%     ~     (p=0.730 n=5+5)
EndpointGeneration/1000/1-8      62.0 ± 0%      62.0 ± 0%     ~     (all equal)

```

The -39% is dramatically understated due to the n^2 behavior of the test. I couldn't get the benchmarks to use more than 100 services as it quickly got so slow the benchmark wouldn't finish. Real cluster dropping from 26 CPUs to 6 CPUs:
![2020-14-07_15-42-03](https://user-images.githubusercontent.com/623453/87483465-a0dacd80-c5e8-11ea-8010-e1da39c897ff.png)