@istio-testing 2020-07-14
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
lint_istio | ff5b13e664400c068967f1a511a754c139fe68ba | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25518/lint_istio/15987) | `/test lint_istio`
unit-tests_istio | ff5b13e664400c068967f1a511a754c139fe68ba | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25518/unit-tests_istio/16605) | `/test unit-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->