@istio-testing 2020-07-14
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
integ-galley-k8s-tests_istio | ff5b13e664400c068967f1a511a754c139fe68ba | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25518/integ-galley-k8s-tests_istio/15975) | `/test integ-galley-k8s-tests_istio`
unit-tests_istio | a8b564117ef969ba6ec2186820e54620b5fef59b | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25518/unit-tests_istio/16611) | `/test unit-tests_istio`
lint_istio | fc8db5d213a92a86e69aabc0985260fb931ea453 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25518/lint_istio/15997) | `/test lint_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->